refactor: moved pages to components/pages
