+++ app/components/pages/error/ErrorPage.js
