+++ app/components/pages/manuscript/ManuscriptPage.js
