+++ app/components/pages/submission-wizard/Author/AuthorDetails.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Action, ErrorText } from '@pubsweet/ui'
-import ValidatedField from '../../ui/atoms/ValidatedField'
+import ValidatedField from '../../../ui/atoms/ValidatedField'
 import TwoColumnRow from '../TwoColumnRow'
 
 const AuthorDetails = ({ fetchOrcid, loading, error }) => (
