+++ app/components/pages/submission-wizard/Author/AuthorDetails.md
