+++ app/components/pages/submission-wizard/Author/AuthorDetailsPage.js
