+++ app/components/pages/submission-wizard/Author/AuthorDetailsSchema.js
