+++ app/components/pages/submission-wizard/Editors/ReviewerSuggestionsSchema.js
