+++ app/components/pages/submission-wizard/Files/FileUploadsSchema.js
