+++ app/components/pages/submission-wizard/ProgressBar.js
