+++ app/components/pages/submission-wizard/Submission/ManuscriptMetadataSchema.js
