+++ app/components/ui/molecules/manuscript/Manuscript.js
