+++ app/components/ui/molecules/manuscript/Manuscript.md
